persist versions to a s3 bucket
having the versions tied to your ci deployment is a very bad idea™ -
when your CI server goes away you don't want your product to start again
at 0.

picking s3 is arbitrary here, but was chosen as it has a fairly low
barrier to entry. git support may be added in the future, though it
seems kind of cludgy to have to commit and push a single number every
time. (the ordering guarantees on the other hand would be nice...)