/out can perform bumps in-place
also pushed handling of initial version into driver, since it's needed
in both Bump and Check. this pulls logic out of the caller as well.

note that the only supported driver right now is still S3, which does
not support compare-and-swaps, but the interface is now fleshed-out to
the point where a Git driver can be dropped in.

[#93919118]