Build golang project in Dockerfile
Change the Dockerfile to build the project in Golang using
golang-alpine container.

The intention is be able to easily compile this resource using
DockerHub. We do not run any testing in the compiling phase.

At the end we delete the code and the go installation, although this
won't affect the image size as most of the space is consume by
underlying layers (golang)