Allow for alternative refs
We store our semver in a file on a different ref than is standard. This allows the repo to store the info, but
for our semver branch to never show up along side other branches unless specifically fetched. There's some
shuffling that needs to happen between the standard branching model and using an alternative ref, but this
change allows you to simply specify another ref and use the specified branch from it (which could also be
"master").

Signed-off-by: Matt Kulka <mkulka@parchment.com>