Fixes concourse/semver-resource#21
This is the smallest change I could make to fix this issue. I'm unable
to run the "unit" tests because I do not have access (or have
personally work with) an OpenStack environment.

The swift driver is actually the only driver with a constructor
(`NewABC()`)  function. The other drivers are constructed in
`FromSource()` where there is a copy of the `if..else` this commit adds.

Signed-off-by: Taylor Silva <tsilva@pivotal.io>